"fixed" a long standing issue with gmake2 generator.
Ok, this commit is a rewrite of @Jarod42's pull request #27. Basically, instead of using file config's `buildcommands`, `buildmessage`, etc. fields, it uses the private `_cfgset` field.
I added a helper to "hide" this workaround, because using premake-core's private stuff is likely to break in the future, and so refactoring at that time will be easier if there is just 1 place to modify.
Using this method seems to remove the need for the `customBakeFiles` function that I was using to insert output files generated by the commands into the project. And the special case of the include directory (which was added in the case of the .ui files) was moved into `customAddFileConfig`.
Also, the QRC's dependencies were changed from relative to absolute, because it appears premake-core handles that for us.
All in all, this change is very handy because it appears to be a lot more integrated into premake-core's inner working, but it comes with the risk of breaking in the future if the private `_cfgset` field changes, thus this long message to explain things. (I don't think `_cfgset` will change anytime soon though, so we should be "safe" for a while)